Move some test-specific functions in gnu.testlet.TestUtils to avoid code duplication
