+++ tests/com/nokia/mid/impl/jms/core/TestLauncher.java
@@ -2,29 +2,13 @@ package com.nokia.mid.impl.jms.core;
 
 import gnu.testlet.TestHarness;
 import gnu.testlet.Testlet;
+import gnu.testlet.TestUtils;
 import java.io.*;
 import javax.microedition.io.file.*;
 import javax.microedition.io.*;
 
 public class TestLauncher implements Testlet {
     native boolean checkImageModalDialog();
-
-    byte[] read(InputStream is) throws IOException {
-        int l = is.available();
-        byte[] buffer = new byte[l+1];
-        int length = 0;
-
-        while ((l = is.read(buffer, length, buffer.length - length)) != -1) {
-            length += l;
-            if (length == buffer.length) {
-                byte[] b = new byte[buffer.length + 4096];
-                System.arraycopy(buffer, 0, b, 0, length);
-                buffer = b;
-            }
-        }
-
-        return buffer;
-    }
 
     public void test(TestHarness th) {
         try {
@@ -34,7 +18,7 @@ public class TestLauncher implements Testlet {
             }
             OutputStream os = file.openDataOutputStream();
             InputStream is = getClass().getResourceAsStream("/org/mozilla/io/test.jpg");
-            os.write(read(is));
+            os.write(TestUtils.read(is));
             os.close();
 
             boolean expectedFailure = false;
