+++ tests/com/nokia/mid/ui/TestTextEditor.java
@@ -2,6 +2,7 @@ package com.nokia.mid.ui;
 
 import gnu.testlet.TestHarness;
 import gnu.testlet.Testlet;
+import gnu.testlet.TestUtils;
 import javax.microedition.lcdui.Alert;
 import javax.microedition.lcdui.Canvas;
 import javax.microedition.lcdui.Display;
@@ -10,71 +11,6 @@ import javax.microedition.lcdui.Graphics;
 import javax.microedition.lcdui.TextField;
 
 public class TestTextEditor extends Canvas implements Testlet {
-    public static String checkCodeFormat(String code) {
-        if (code == null) {
-            return code;
-        }
-        if (code.length() == 10) {
-            String c1 = code.substring(0, 5);
-            String c2 = code.substring(5);
-            return c1 + " " + c2;
-        } else if (code.length() == 6) {
-            String c1 = code.substring(0, 2);
-            String c2 = code.substring(2);
-            return c1 + " " + c2;
-        } else {
-            return code;
-        }
-    }
-
-    private final static int LEAD_OFFSET = 0xD800 - (0x10000 >> 10);
-
-    public static String getSurrogatePairs(String inputString) {
-        int character;
-        char low, high;
-        int start = 0, end = 0;
-        if (inputString == null) {
-            return inputString;
-        }
-
-        StringBuffer sb = new StringBuffer(1000);
-
-        // Go through all characters in the input.
-        // Space (0x20) is used as separator
-        do {
-            end = inputString.indexOf(" ", start);
-
-            // Space not found -> last sub-string
-            if (end == -1) {
-                end = inputString.length();
-            }
-
-            try {
-                character = Integer.parseInt(inputString.substring(start, end), 16);
-            } catch (Exception e) {
-                e.printStackTrace();
-                return null;
-            }
-
-            // Anything below 0xffff is not surrogate pair
-            if (character < 0xffff) {
-                sb.append((char) character);
-            } else {
-                // From http://www.unicode.org/faq/utf_bom.html
-                high = (char) (LEAD_OFFSET + (character >> 10));
-                low = (char) (0xDC00 + (character & 0x3FF));
-
-                sb.append(high);
-                sb.append(low);
-            }
-
-            // skip the space
-            start = (end + 1);
-        } while (end != inputString.length());
-
-        return sb.toString();
-    }
-
     public void testConstraints(TestHarness th, int constraints, int tolerance) {
         TextEditor textEditor = new TextEditor("Hello, world!", 20, 0, 100, 24);
 
@@ -199,7 +135,7 @@ public class TestTextEditor extends Canvas implements Testlet {
 
         th.check(textEditor.getCaretPosition(), 13);
 
-        String code = getSurrogatePairs(checkCodeFormat("1f609"));
+        String code = TestUtils.getEmojiString("1f609");
         textEditor.insert(code, textEditor.getCaretPosition());
 
         th.check(textEditor.getCaretPosition(), 14);
@@ -277,29 +213,29 @@ public class TestTextEditor extends Canvas implements Testlet {
         // emoji, we allow removing one of the two codepoints with the |delete| method and
         // we consider one emoji with two codepoints as taking two spaces in the TextEditor.
         // Hence, we're testing this behavior here.
-        textEditor.setContent(getSurrogatePairs(checkCodeFormat("1f1ee1f1f9")));
+        textEditor.setContent(TestUtils.getEmojiString("1f1ee1f1f9"));
         th.check(textEditor.getCaretPosition(), 2);
         th.check(textEditor.size(), 2);
         textEditor.delete(1, 1);
-        th.check(textEditor.getContent(), getSurrogatePairs(checkCodeFormat("1f1ee")));
+        th.check(textEditor.getContent(), TestUtils.getEmojiString("1f1ee"));
         th.check(textEditor.getCaretPosition(), 1);
         th.check(textEditor.size(), 1);
-        textEditor.setContent(getSurrogatePairs(checkCodeFormat("1f1ee1f1f9")));
+        textEditor.setContent(TestUtils.getEmojiString("1f1ee1f1f9"));
         textEditor.delete(0, 1);
-        th.check(textEditor.getContent(), getSurrogatePairs(checkCodeFormat("1f1f9")));
+        th.check(textEditor.getContent(), TestUtils.getEmojiString("1f1f9"));
         th.check(textEditor.getCaretPosition(), 0);
         th.check(textEditor.size(), 1);
 
-        textEditor.setContent(getSurrogatePairs(checkCodeFormat("2320e3")));
+        textEditor.setContent(TestUtils.getEmojiString("2320e3"));
         th.check(textEditor.getCaretPosition(), 2);
         th.check(textEditor.size(), 2);
         textEditor.delete(1, 1);
         th.check(textEditor.getContent(), "#");
         th.check(textEditor.getCaretPosition(), 1);
         th.check(textEditor.size(), 1);
-        textEditor.setContent(getSurrogatePairs(checkCodeFormat("2320e3")));
+        textEditor.setContent(TestUtils.getEmojiString("2320e3"));
         textEditor.delete(0, 1);
-        th.check(textEditor.getContent(), getSurrogatePairs(checkCodeFormat("20e3")));
+        th.check(textEditor.getContent(), TestUtils.getEmojiString("20e3"));
         th.check(textEditor.getCaretPosition(), 0);
         th.check(textEditor.size(), 1);
