+++ tests/gfx/TestLauncher.java
@@ -4,26 +4,10 @@ import java.io.*;
 import javax.microedition.io.file.*;
 import javax.microedition.io.*;
 import javax.microedition.midlet.*;
-import com.nokia.mid.impl.jms.core.*;
+import com.nokia.mid.impl.jms.core.Launcher;
+import gnu.testlet.TestUtils;
 
 public class TestLauncher extends MIDlet {
-    byte[] read(InputStream is) throws IOException {
-        int l = is.available();
-        byte[] buffer = new byte[l+1];
-        int length = 0;
-
-        while ((l = is.read(buffer, length, buffer.length - length)) != -1) {
-            length += l;
-            if (length == buffer.length) {
-                byte[] b = new byte[buffer.length + 4096];
-                System.arraycopy(buffer, 0, b, 0, length);
-                buffer = b;
-            }
-        }
-
-        return buffer;
-    }
-
     public void startApp() {
         try {
             FileConnection file = (FileConnection)Connector.open("file:////test.png", Connector.READ_WRITE);
@@ -32,7 +16,7 @@ public class TestLauncher extends MIDlet {
             }
             OutputStream os = file.openDataOutputStream();
             InputStream is = getClass().getResourceAsStream("/gfx/images/red.png");
-            os.write(read(is));
+            os.write(TestUtils.read(is));
             os.close();
 
             Launcher.handleContent("test.png");
