+++ tests/javax/microedition/media/TestMediaImage.java
@@ -2,6 +2,7 @@ package javax.microedition.media;
 
 import gnu.testlet.TestHarness;
 import gnu.testlet.Testlet;
+import gnu.testlet.TestUtils;
 import javax.microedition.lcdui.*;
 import javax.microedition.midlet.*;
 import java.io.*;
@@ -12,23 +13,6 @@ import javax.microedition.media.control.*;
 
 public class TestMediaImage implements Testlet, PlayerListener {
     TestHarness th;
-
-    byte[] read(InputStream is) throws IOException {
-        int l = is.available();
-        byte[] buffer = new byte[l+1];
-        int length = 0;
-
-        while ((l = is.read(buffer, length, buffer.length - length)) != -1) {
-            length += l;
-            if (length == buffer.length) {
-                byte[] b = new byte[buffer.length + 4096];
-                System.arraycopy(buffer, 0, b, 0, length);
-                buffer = b;
-            }
-        }
-
-        return buffer;
-    }
 
     public void test(TestHarness th) {
         this.th = th;
@@ -42,7 +26,7 @@ public class TestMediaImage implements Testlet, PlayerListener {
             }
             OutputStream os = file.openDataOutputStream();
             InputStream is = getClass().getResourceAsStream("/org/mozilla/io/test.jpg");
-            os.write(read(is));
+            os.write(TestUtils.read(is));
             os.close();
 
             Player player = Manager.createPlayer("file:////test.jpg");
