+++ tests/org/mozilla/io/TestNokiaImageProcessingInMultiThread.java
@@ -9,6 +9,7 @@ import com.nokia.mid.s40.codec.DataEncoder;
 import com.nokia.mid.s40.codec.DataDecoder;
 import gnu.testlet.Testlet;
 import gnu.testlet.TestHarness;
+import gnu.testlet.TestUtils;
 
 public class TestNokiaImageProcessingInMultiThread implements Testlet {
     LocalMessageProtocolConnection client;
@@ -23,23 +24,6 @@ public class TestNokiaImageProcessingInMultiThread implements Testlet {
             this.th = th;
         }
 
-        byte[] read(InputStream is) throws IOException {
-            int l = is.available();
-            byte[] buffer = new byte[l+1];
-            int length = 0;
-
-            while ((l = is.read(buffer, length, buffer.length - length)) != -1) {
-                length += l;
-                if (length == buffer.length) {
-                    byte[] b = new byte[buffer.length + 4096];
-                    System.arraycopy(buffer, 0, b, 0, length);
-                    buffer = b;
-                }
-            }
-
-            return buffer;
-        }
-
         public void run() {
             try {
                 LocalMessageProtocolConnection client = (LocalMessageProtocolConnection)Connector.open("localmsg://" + PROTO_NAME);
@@ -49,7 +33,7 @@ public class TestNokiaImageProcessingInMultiThread implements Testlet {
                 }
                 OutputStream os = originalImage.openDataOutputStream();
                 InputStream is = TestNokiaImageProcessingInMultiThread.class.getResourceAsStream("test.jpg");
-                os.write(read(is));
+                os.write(TestUtils.read(is));
                 os.close();
 
                 DataEncoder dataEncoder = new DataEncoder("Conv-BEB");
